requirements.txt
flask>=1.1.2
pillow>=8.1.2
brother_ql
